Made the Play button functional!
Now if the user clicks on the Play button he will be taken to another pygame window he can play the game.  Now there is a moving basket as well as randomly generated eggs to be collected.  Once the user presses either the Right or Left arrow key on the keyboard the basket starts to move either left or right with respect to the keys being pressed by shifting 7 pixels.  But it won't stop which has to be fixed.